Move folders package into libs
